[uss_qualifier] move synchronization code to subpackage
