+++ monitoring/uss_qualifier/scenarios/astm/utm/dss/__init__.py
@@ -1,3 +1,2 @@
 from .subscription_validation import SubscriptionValidation
 from .subscription_simple import SubscriptionSimple
-from .subscription_synchronization import SubscriptionSynchronization
