+++ monitoring/uss_qualifier/suites/astm/utm/dss_probing.yaml
@@ -35,7 +35,7 @@ actions:
         all_dss_instances: all_dss_instances
         planning_area: planning_area
   - test_scenario:
-      scenario_type: scenarios.astm.utm.dss.SubscriptionSynchronization
+      scenario_type: scenarios.astm.utm.dss.synchronization.SubscriptionSynchronization
       resources:
         dss: dss
         other_instances: all_dss_instances
